Enable the syncronized HighLighting of the image and the text. A button on the top of the image panel switch on and off the syncronization. Each zone/line can be selected and deselected by clicking on it (line or zone). A new property is added in ui_config to activate the highlight syncronization button
